Break apart Admin/AdminContainer and App/AppContainer into separate modules
