+++ client/js/components/app-container.js
@@ -1,0 +1,8 @@
+var React = require("react");
+var connect = require("react-redux").connect;
+
+var App = require("./app"); 
+
+var AppContainer = connect()(App);
+
+module.exports = AppContainer;
